+++ src/modulos/posicionamineto/Desplazar.js
@@ -1,18 +1,19 @@
-import $ from "jquery";
-import Posicionamineto from "./Posicionamiento";
+
+import Posicionamiento from "./Posicionamiento";
 
 
 class Desplazar {
-    static ejecutar(origen, ele, metodoPreguntar,metodoPosicionar, mueca = false, clase) {
-        if(metodoPreguntar(origen, ele)) {
-            $(ele).css("left",Posicionamineto.reacomodamientoHorizontal(origen,ele))
-            $(ele).css("top",Posicionamineto.reacomodamientoVertical(origen,ele))
-            if(mueca) {
-                let m = $("<span class="+ clase + "></span>")
-                $(ele).append(m)
+    static ejecutar(origen, ele, metodoPreguntar, metodoPosicionar, mueca = false, clase) {
+        if (metodoPreguntar(origen, ele)) {
+            ele.style.left = Posicionamiento.reacomodamientoHorizontal(origen, ele) + "px"
+            ele.style.top = Posicionamiento.reacomodamientoVertical(origen, ele) + "px"
+            if (mueca) {
+                let m = document.createElement("span")
+                m.classList.add(clase)
+                ele.appendChild(m)
             }
             metodoPosicionar(origen, ele)
-            return true 
+            return true
         }
 
         return false